requirements.txt
astroid==3.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mysql-connector-python==8.3.0
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.1.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
tomlkit==0.12.4
typing_extensions==4.10.0
uuid==1.30
Werkzeug==3.0.1
Flask-WTF
mysql-connector-python
